backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow-cpu==2.16.1
keras==3.13.0
numpy==1.26.4
pillow==10.2.0
python-dotenv==1.0.0
flask-jwt-extended==4.6.0
bcrypt==4.2.1
mysql-connector-python==9.2.0
